Split Quickstarts stage to PR and tag and reset
Move reset task to the separate step to identify the problem faster

The problem is:
If reset falls - it fails the promote job. If promote job fails then snapshots doesnt gets updated. In this case kogito pipeline could not proceed.

Currently need to separate update of snapshots from the promote part and override tag if it is already exist so we can rerun promote job.